% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CqCalc.R
\name{calc.Cq.qpcR}
\alias{calc.Cq.qpcR}
\title{calculates Cq values}
\usage{
calc.Cq.qpcR(method = "TP", fct = l4, cq.new = "calc.qpcR.Cq")
}
\arguments{
\item{method}{The method used to calculate the Cq value. "TP / cpD1" for Turning Point (first derivative) . "SD / cpD2" for first exponential incline (second derivative). "cpE", "cpR", "cpT", "Cy0", "cpCQ", "cpMR" (see qpcR package)}

\item{fct}{will be used by qpcR as fitting function. l4, l5, b4}

\item{cq.new}{The Column name of the new Cq values.}
}
\value{
does not return! adds Cq value to input.cq data frame
}
\description{
WRAPPER for qpcR Package
'qpcR: Modelling and Analysis of Real-Time PCR Data, Andrej-Nikolai Spiess, 2018'
}
\details{
This function will calculate Cq values for the data frame input.cq
You need to have input.cq and input.raw from "read.fluorescenceTable" / "read.cqTable" or build them manually.

Please note, that it is faster to use qpcR directly!
}
