% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CqCalc.R
\name{calcCq}
\alias{calcCq}
\title{calculates Cq values}
\usage{
calcCq(method = "TP", fct = "ll5", cq.new = "calcCq")
}
\arguments{
\item{method}{The method used to calculate the Cq value. "TP" for Turning Point (first derivative) . "SD" for first expnential incline (second derivative).}

\item{fct}{will be used by {drc} as fitting function. "ll4" or "ll5" is implemented.}

\item{cq.new}{The Column name of the new Cq values.}
}
\value{
does not return! adds Cq value to input.cq data frame
}
\description{
This function will calculate Cq values for the data frame input.cq
You need to have input.cq and input.raw and input.raw.meld from "read.fluorescenceTable" / "read.cqTable" or build them manuaally.
}
\details{
!This will fail sometimes, do to uniroot() function returning 0!
}
