% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCqData.R
\name{delta.Cq.data}
\alias{delta.Cq.data}
\title{This Function will give the delta Cq values. Always: offtarget Cq - target Cq
\itemize{
\item helper function.
}}
\usage{
delta.Cq.data(
  CqType = NA,
  method = "combinatorial",
  onlyNumeric = FALSE,
  return = FALSE
)
}
\arguments{
\item{CqType}{this is the Cq value columns from the input.cq that should be used.}

\item{method}{can be either "combinatorial", "random", "direct" or "mean" (can be abbriviated). This will impact how the differences are calculated.}

\item{onlyNumeric}{Will only use samples that are a numerical (for percentages the "\%" will be stripped)}

\item{return}{standard=FALSE will write in global scope! Otherwise will return the dataframe.}
}
\value{
standard: returns nothing. Creates a dataframe or creates a a dataframe data.cq.sum in global scope
}
\description{
The delta Cq values can be calculated by randomly assign two values, build all combinations of deta Cq. Directly use the values from the same row or just use the mean values.
}
