% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCqData.R
\name{linSqrtTransform}
\alias{linSqrtTransform}
\title{This Function will linearise the Delta Cq data with a shifted square root transformation.
\itemize{
\item helper function.
}}
\usage{
linSqrtTransform(data)
}
\arguments{
\item{data}{A data frame wich has sample and delta.Cq column (delta.Cq) But transformation will only affect sample column!}
}
\value{
returnes a dataframe
}
\description{
This Function will linearise the Delta Cq data with a shifted square root transformation.
\itemize{
\item helper function.
}
}
