% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCqData.R
\name{make.Cq.data}
\alias{make.Cq.data}
\title{This Function will create the main Data object for all analysis.
Used after an input is created with read.CqTable and all optional calculation data validation is complete.}
\usage{
make.Cq.data(
  add = FALSE,
  target = "Genotype A",
  CqType = c("TP", "SD"),
  outliers = TRUE,
  alpha = 0.05,
  outlier.range = 3,
  silent = FALSE
)
}
\arguments{
\item{add}{This toggle wil add the samples, if an data.cq is already existing in global scope.}

\item{target}{the target genotype "genotype A".}

\item{CqType}{this is the Cq value columns from the input.cq that should be used.}

\item{outliers}{logical if outliers are to be deleted from the output}

\item{alpha}{alpha for outlier testing (0.05 = 95\% significance)}

\item{outlier.range}{This is only important for samples with 3 or less values. In this case the range of data (e.g. Range c(1,1.4,1.3) = 0.4) need to be at least outlier.range if an outlier test should happen. Normally outlier test for 3 or less values is not recommended. But this helps to get rid of clear outliers e.g. (2,2,30). My advice is to check the data also manually.}

\item{silent}{If status of outlier detection and processing is printed.}
}
\value{
returns a list of samples with cq values (data.cq)
}
\description{
Outliers can be removed. To check the results one can use the table.Cq() function before.
}
\details{
If an data.Cq object already exists, it will be overwritten when add = FALSE.
Otherwise samples will be added. Or overwritten! It is not jet possible to add more values in a sample...
}
