% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CqPlots.R
\name{plot.Cq.efficiency}
\alias{plot.Cq.efficiency}
\title{plots Cq values of each genotype againgst the pseudo log concentration.}
\usage{
plot.Cq.efficiency(CqType = "SD")
}
\arguments{
\item{CqType}{this is the Cq value type that should be used.}
}
\value{
returns a ggplot with the Cq values for both genotypes.
}
\description{
This function will give you a classic efficiency plot. Therefore the Cq values are plotted.
You need to have data.cq list from make.cq.data() function.
}
\details{
! The samplename is the concentration / mixture for the target genotype!
Only valid numeric samples will be used. Sample names like 100, "100", "100\%" will be converted to 100. But other sample names like "Test" will be ignored.
}
