% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CqPlots.R
\name{plot.Cq}
\alias{plot.Cq}
\title{plots Cq values with the both genotypes on x and y axis}
\usage{
plot.Cq(CqType = "Regression", onlyNumeric = FALSE)
}
\arguments{
\item{CqType}{this is the Cq value type that should be used.}

\item{onlyNumeric}{Will only use samples that are a numerical (e.g. percentage)}
}
\value{
returns a ggplot with the Cq values for both genotypes.
}
\description{
This function will give you a Cq plot. Therefore the Cq values are plotted.
You need to have data.cq list from make.cq.data() function.
}
