% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCurves.R
\name{plot.curve}
\alias{plot.curve}
\title{plot fluorescence curves}
\usage{
plot.curve(
  sample = "all",
  target = "Genotype A",
  color.target = "firebrick2",
  color.offtarget = "cornflowerblue"
)
}
\arguments{
\item{sample}{specify the sample from input.cq std: "all" will plot all curves. Can be Numeric e.g. 100 for dilution series etc. Can be Character e.g. "100" for a normal experiment.}

\item{target}{the target genotype "genotype A".}

\item{color.target}{The color for the target genotype curve (only if not all curves are plotted)}

\item{color.offtarget}{The color for the other genotype curve (only if not all curves are plotted)}
}
\value{
a ggplot object
}
\description{
plot fluorescence curves
}
