% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaCqPlots.R
\name{plot.delta.Cq.differences}
\alias{plot.delta.Cq.differences}
\title{plots delta Cq values against the square root of the concentration}
\usage{
plot.delta.Cq.differences(
  CqType = c("SD", "TP"),
  method = "c",
  xlab = "mixture [\%] target of interest"
)
}
\arguments{
\item{CqType}{list of Cq types that should be plotted.}

\item{method}{to generate the delta Cq values. See delta.Cq.data()}

\item{xlab}{x axis title for the plots.}
}
\value{
returns a R base plot.
}
\description{
will use the data.Cq list as input.
Only numeric values will be used (100 = 100 \% = "100")
}
