% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetaCqPlots.R
\name{plot.delta.Cq.overview}
\alias{plot.delta.Cq.overview}
\title{plots delta Cq values against the concentration}
\usage{
plot.delta.Cq.overview(
  CqType = c("SD", "TP"),
  linSqrtTrans = FALSE,
  method = "c",
  xlab = "mixture [\%]"
)
}
\arguments{
\item{CqType}{list of Cq types that should be plotted.}

\item{linSqrtTrans}{Will transform the values to linearise the values! this is basically a shifted square root representation.}

\item{method}{to generate the delta Cq values. See delta.Cq.data()}

\item{xlab}{x axis title for the plots.}
}
\value{
returns a R base plot.
}
\description{
will use the data.Cq list as input.
Only numeric values will be used (100 = 100 \% = "100")
}
