% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readqPCR.R
\name{read.cqTable}
\alias{read.cqTable}
\title{Load cq table}
\usage{
read.cqTable(csv, sep = ",", dec = ",")
}
\arguments{
\item{csv}{Path to the input csv table}

\item{sep}{cell separator "," "." or ...}

\item{dec}{decimals as "," or "." or ...}
}
\value{
returns a data frame of cq values.
}
\description{
This function is a wrapper for read.csv
#'
}
