% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readqPCR.R
\name{read.fluorescenceTable}
\alias{read.fluorescenceTable}
\title{Load fluorescence table}
\usage{
read.fluorescenceTable(csv, sep = ",", dec = ",")
}
\arguments{
\item{csv}{Path to the input csv table}

\item{sep}{cell separator "," "." or ...}

\item{dec}{decimals as "," or "." or ...}
}
\value{
does not return! input.raw and input.raw.melt data frames will be saved to global.
}
\description{
This function is a wrapper for read.csv
The input.raw and input.raw.melt data frames will be saved to global scope.
Rigth now you only can have 1 table for 1 analyis. You can consider to add more values manually or combine the csv files beforehand. Make sure the wells are named individually.
The first column need to be called "Cycle"!
}
