% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regression.R
\name{regression.delta.Cq}
\alias{regression.delta.Cq}
\title{This function wil perform a regression analysis on the delta Cq values.}
\usage{
regression.delta.Cq(
  CqType = "SD",
  method = "c",
  fit = "linear",
  rawPolynomials = FALSE,
  cv = FALSE,
  cv.seed = sample(1:100, 1),
  cv.method = "LGOCV",
  cv.p = 50
)
}
\arguments{
\item{CqType}{this is the Cq value type that should be used.}

\item{method}{method for generating the delta Cq values (see delta.Cq.data())}

\item{fit}{model for lm() function. "linear", "poly3", "poly4".}

\item{cv}{should a cross validation be made? With caret!}

\item{cv.seed}{seed for cv}

\item{cv.method}{The method for cv from caret package}

\item{cv.p}{percentage of training data}
}
\value{
returns a model: model.delta.Cq object in global space.
}
\description{
The model and prediction with investR package are reported. A CrossValidation for the model is possible with caret.
The cross validation here is a basic wrapper for caret package. If more detailed CV is wanted, the analysis should be made seperatly with caret.
}
