% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaCqCalc.R
\name{table.Cq}
\alias{table.Cq}
\title{shows Cq values from a sample and the mean Delta Cq}
\usage{
table.Cq(
  sample = NA,
  target = "Genotype A",
  CqType = c("TP", "SD"),
  outliers = TRUE,
  alpha = 0.05,
  outlier.range = 3,
  decimals = 3,
  format = "kable",
  silent = FALSE
)
}
\arguments{
\item{sample}{specify the sample from input.cq}

\item{target}{the target genotype "genotype A".}

\item{CqType}{wich Cq values should be used. This can be a vector!}

\item{outliers}{logical if outliers are to be deleted from the output}

\item{alpha}{alpha for outlier testing (0.05 = 95\% significance)}

\item{outlier.range}{This is only important for samples with 3 or less values. In this case the range of data (e.g. Range c(1,1.4,1.3) = 0.4) need to be at least outlier.range if an outlier test shoud happen. Normally outlier test for 3 or less values is not recommended. But this helps to get rid of clear outliers e.g. (2,2,30). My advice is to check the data also manually.}

\item{decimals}{decimals for the resulting table}

\item{format}{How the table will be formated. possible are "kable" and "DT" (work in progress) or "data" for pure dataframe, "fulldata" for a dataframe with difference, mean, and sd. Or any other input will give just the table.}

\item{silent}{should status be printed? (mostly for outlier detection)}
}
\value{
returns a table with the delta Cq values with mean and standard deviation.
}
\description{
This function will show the Cq values from the samples.
You need to have input.cq "read.cqTable" or build it manually with Cq values present.
}
