% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{voges_grubbs}
\alias{voges_grubbs}
\title{Helper function for outlier detection.}
\usage{
voges_grubbs(data, outlier.range = 6, alpha = 0.05, silent = FALSE)
}
\arguments{
\item{data}{input data (List, dataframe or vector?)}

\item{outlier.range}{How many values have to be present for doing grubbs test (std 6)}

\item{alpha}{p-value for outlier removal (0.05 = 95\% significance)}

\item{silent}{should outliers be printed to output?}
}
\value{
returns a dataframe or list depending on the input.
}
\description{
Wrapper for the Grubbs test from outliers package. This test will scan for multiple outliers! removing one after another.
}
