% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetaCqPlots.R
\name{plot.delta.Cq.box}
\alias{plot.delta.Cq.box}
\title{plots delta Cq values as box-plots}
\usage{
plot.delta.Cq.box()
}
\value{
returns a ggplot object.
}
\description{
delta.Cq data frame has to be created before
}
