% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaCqPlots.R
\name{plot.delta.Cq.differences}
\alias{plot.delta.Cq.differences}
\title{plots delta Cq values against the square root of the concentration}
\usage{
plot.delta.Cq.differences(
  xlab = "Percent of target genotype [\%]",
  breaks = c(0, 5, 10, 20, 30, 40, 50)
)
}
\arguments{
\item{xlab}{x axis title for the plots.}

\item{breaks}{list of x ticks / breaks}
}
\value{
returns a R base plot.
}
\description{
delta.Cq data frame has to be created before.
Used the mean() and sd() base functions.
}
