% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readqPCR.R
\name{read.fluorescenceTable}
\alias{read.fluorescenceTable}
\title{Load fluorescence table}
\usage{
read.fluorescenceTable(csv, sep = ",", dec = ",")
}
\arguments{
\item{csv}{Path to the input csv table}

\item{sep}{cell separator "," "." or ...}

\item{dec}{decimals as "." or "," or ...}
}
\value{
returns a data frame of fluorescence values.
}
\description{
This function is a wrapper for read.csv
#'
}
