% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCqData.R
\name{reTransform}
\alias{reTransform}
\title{This Function will retransform a value / list etc to normal percentage / proportion}
\usage{
reTransform(value)
}
\arguments{
\item{value}{A signle value that should be retransformed.}
}
\value{
returnes a dataframe
}
\description{
The value(s) needs to be between -sqrt(50) and sqrt(50) or the returned value will not be correct.
}
\details{
\itemize{
\item helper function.
}
}
